feat: remove the check for existing apps
We use the Job ID in our containerID value, and the Job ID will be
unique for every job run, so there should never be a circumstance where
an existing app will be found. If it didn't hurt to check I would keep
this, but the call to cloud.gov is a waste.